dwarves: Switch type_id_t from uint16_t to uint32_t
With BTF dedup we end up with a CU with all the types, and with
something like an allmodconfig vmlinux image it ends up with more than
65535 types, so use uint32_t for the type IDs to accomodate that.

Reported-by: Andrii Nakryiko <andriin@fb.com>
Tested-by:Andrii Nakryiko <andrii.nakryiko@gmail.com>
Link: https://lore.kernel.org/bpf/CAEf4Bzb0SpvXdDKMMnUof==kp4Y0AP54bKFjeCzX_AsmDm7k7g@mail.gmail.com/

+++ dwarves.h
@@ -27,7 +27,11 @@ enum load_steal_kind {
 	LSK__STOP_LOADING,
 };
 
-typedef uint16_t type_id_t;
+/*
+ * BTF combines all the types into one big CU using btf_dedup(), so for something
+ * like a allyesconfig vmlinux kernel we can get over 65535 types.
+ */
+typedef uint32_t type_id_t;
 
 /** struct conf_load - load configuration
  * @extra_dbg_info - keep original debugging format extra info
